refactor(auth): extract a shared field-error helper
login.js and register.js carried an identical showError body (toggling
aria-invalid on the fields plus a localized message lookup with a generic
fallback). Move that logic into ui/form-error.js. Both pages keep their
own local signature, so no call site changes.
