+++ public/js/login.js
@@ -6,6 +6,7 @@ import { login, me, changePassword, getPasswordPolicy, registrationEnabled } fro
 import { initI18n, applyI18n, t, fmtDate } from './core/i18n.js';
 import { bindPasswordStrength } from './ui/password-strength.js';
 import { mountFloatingBackground } from './ui/floating-bg.js';
+import { applyFieldError } from './ui/form-error.js';
 
 const $ = (id) => document.getElementById(id);
 
@@ -22,19 +23,7 @@ const ERROR_FIELDS = {
 };
 
 function showError(scope, code, extra = {}) {
-  const el = document.getElementById(`${scope}-error`);
-  if (!el) return;
-  const invalid = !!code;
-  for (const id of ERROR_FIELDS[scope] || []) {
-    const input = document.getElementById(id);
-    if (!input) continue;
-    if (invalid) input.setAttribute('aria-invalid', 'true');
-    else input.removeAttribute('aria-invalid');
-  }
-  if (!code) { el.textContent = ''; return; }
-  const key = code.includes('.') ? code : `errors.${code}`;
-  const fallback = t('errors.generic');
-  el.textContent = t(key, extra) === key ? fallback : t(key, extra);
+  applyFieldError(`${scope}-error`, ERROR_FIELDS[scope] || [], code, extra);
 }
 
 function redirectAfterAuth(user) {
